Refactor: Move save_forecast function to save.py
